Saperate User and Post to models, create folder flaskpro import forms models there and create init files. rename main.py run.py move all route to routes.py saperate with db configuration make all project much more oriented structure
