+++ flask_pro/flaskpro/static/bootstrap/js/ie10-viewport-bug-workaround.js
