+++ flask_pro/flaskpro/static/images/student.ico
