+++ flask_pro/flaskpro/templates/navbar.html
@@ -17,6 +17,7 @@
 	                <li class="active"><a href="/home">Home</a></li>
 	                <li><a href="/about">About</a></li>
 	                <li><a href="/contact">Contact</a></li>
+	                <li><a href="/posts">Posts</a></li>
 	                <li class="dropdown">
 	                  <a href="" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">User <span class="caret"></span></a>
 	                  <ul class="dropdown-menu">
