+++ flask_pro/info.txt
@@ -107,3 +107,18 @@ User('dilmac', 'dilmac@gmail.com', 'default.jpg')
 
 Delete all db
 >>> db.drop_all()
+
+# After db created in the terminal show some waring from SQLAlchemy like that:
+flask run
+ * Serving Flask app "main.py"
+ * Environment: production
+   WARNING: Do not use the development server in a production environment.
+   Use a production WSGI server instead.
+ * Debug mode: off
+/home/dilmac/Develop/flaskenv/lib/python3.6/site-packages/flask_sqlalchemy/__init__.py:794: FSADeprecationWarning: SQLALCHEMY_TRACK_MODIFICATIONS adds significant overhead and will be disabled by default in the future.  Set it to True or False to suppress this warning.
+  'SQLALCHEMY_TRACK_MODIFICATIONS adds significant overhead and '
+ * Running on http://127.0.0.1:5000/ (Press CTRL+C to quit)
+
+To stop this warining add the line after app.config['SQLALCHEMY_DATABASE_URI']
+
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
