+++ flask_pro/run.py
@@ -1,0 +1,5 @@
+from flaskpro import app
+
+
+if __name__ == '__main__':
+    app.run(debug=True)
